+++ app/commands/characters_context/pathfinder2/update_command.rb
@@ -143,7 +143,9 @@ module CharactersContext
       end
       # rubocop: enable Metrics/AbcSize, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
 
-      def upload_avatar(input)
+      def upload_avatar(input) # rubocop: disable Metrics/AbcSize
+        return if input.slice(:avatar_file, :avatar_url, :file).keys.blank?
+
         attach_avatar_by_file.call({ character: input[:character], file: input[:avatar_file] }) if input[:avatar_file]
         attach_avatar_by_url.call({ character: input[:character], url: input[:avatar_url] }) if input[:avatar_url]
         input[:character].avatar.attach(input[:file]) if input[:file]